Preliminary changes in media_import.c / adding gf_import_matroska_video.
Changes in media_import.c to accept mkv files.
gf_import_matroska_video added to media_import.c. The code for this
routine is primarily an adjustment of gf_import_avi_video but instead
of AVI lib it uses libavcodec from FFmpeg. For now this routine
extracts some information from the video file and also extracts all the
frames and put them in (char *frame) variable.
It is important to note that the written code is expected to work for
any video container supported by FFmpeg, not only Matroska.